+++ src/trace.rs
@@ -16,13 +16,12 @@ use opentelemetry::{
 };
 use opentelemetry_http::HttpClient;
 use opentelemetry_sdk::{
-    export::trace::{ExportResult, SpanData, SpanExporter},
+    error::OTelSdkResult,
+    trace::{SpanData, SpanExporter},
     Resource,
 };
 use opentelemetry_semantic_conventions as semcov;
-use std::{borrow::Cow, collections::HashMap, future::Future, pin::Pin, sync::Arc, time::Duration};
-
-type BoxFuture<'a, T> = Pin<Box<dyn Future<Output = T> + Send + 'a>>;
+use std::{borrow::Cow, collections::HashMap, sync::Arc, time::Duration};
 
 /// Deprecated semantic convention key for HTTP host
 ///
@@ -98,18 +97,23 @@ impl<C> Exporter<C> {
             data: Some(data),
         });
 
+        let event_resource = if self.resource_attributes_in_events_and_logs {
+            Some(resource)
+        } else {
+            None
+        };
         for event in span.events.iter() {
             let (data, name) = match event.name.as_ref() {
                 x if x == EVENT_NAME_CUSTOM => (
-                    Data::Event(event.into()),
+                    Data::Event(EventAndResource(event, event_resource).into()),
                     "Microsoft.ApplicationInsights.Event",
                 ),
                 x if x == EVENT_NAME_EXCEPTION => (
-                    Data::Exception(event.into()),
+                    Data::Exception(EventAndResource(event, event_resource).into()),
                     "Microsoft.ApplicationInsights.Exception",
                 ),
                 _ => (
-                    Data::Message(event.into()),
+                    Data::Message(EventAndResource(event, event_resource).into()),
                     "Microsoft.ApplicationInsights.Message",
                 ),
             };
@@ -133,18 +137,22 @@ where
     C: HttpClient + 'static,
 {
     /// Export spans to Application Insights
-    fn export(&mut self, batch: Vec<SpanData>) -> BoxFuture<'static, ExportResult> {
+    async fn export(&self, batch: Vec<SpanData>) -> OTelSdkResult {
         let client = Arc::clone(&self.client);
-        let endpoint = Arc::clone(&self.endpoint);
+        let endpoint = Arc::clone(&self.track_endpoint);
         let envelopes: Vec<_> = batch
             .into_iter()
             .flat_map(|span| self.create_envelopes_for_span(span, &self.resource))
             .collect();
 
-        Box::pin(async move {
-            crate::uploader::send(client.as_ref(), endpoint.as_ref(), envelopes).await?;
-            Ok(())
-        })
+        crate::uploader::send(
+            client.as_ref(),
+            endpoint.as_ref(),
+            envelopes,
+            self.retry_notify.clone(),
+        )
+        .await
+        .map_err(Into::into)
     }
 
     fn set_resource(&mut self, resource: &Resource) {
@@ -215,9 +223,7 @@ pub(crate) fn is_remote_dependency_success(span: &SpanData) -> Option<bool> {
 struct SpanAndResource<'a>(&'a SpanData, &'a Resource);
 
 impl<'a> From<SpanAndResource<'a>> for RequestData {
-    fn from(s: SpanAndResource<'a>) -> RequestData {
-        let span = s.0;
-        let resource = s.1;
+    fn from(SpanAndResource(span, resource): SpanAndResource<'a>) -> RequestData {
         let mut data = RequestData {
             ver: 2,
             id: span.span_context.span_id().to_string().into(),
@@ -311,9 +317,7 @@ impl<'a> From<SpanAndResource<'a>> for RequestData {
 }
 
 impl<'a> From<SpanAndResource<'a>> for RemoteDependencyData {
-    fn from(s: SpanAndResource<'a>) -> RemoteDependencyData {
-        let span = s.0;
-        let resource = s.1;
+    fn from(SpanAndResource(span, resource): SpanAndResource<'a>) -> RemoteDependencyData {
         let mut data = RemoteDependencyData {
             ver: 2,
             id: Some(span.span_context.span_id().to_string().into()),
@@ -422,9 +426,14 @@ impl<'a> From<SpanAndResource<'a>> for RemoteDependencyData {
 
         if span.span_kind == SpanKind::Internal {
             data.type_ = Some("InProc".into());
-        } else if let Some(&db_system) = attrs.get(semcov::trace::DB_SYSTEM) {
+        } else if let Some(&db_system) = attrs.get(semcov::trace::DB_SYSTEM_NAME).or_else(|| {
+            attrs.get(
+                #[allow(deprecated)]
+                semcov::attribute::DB_SYSTEM,
+            )
+        }) {
             data.type_ = Some(db_system.into());
-        } else if let Some(&messaging_system) = attrs.get(semcov::trace::MESSAGING_SYSTEM) {
+        } else if let Some(&messaging_system) = attrs.get(semcov::attribute::MESSAGING_SYSTEM) {
             data.type_ = Some(messaging_system.into());
         } else if let Some(&rpc_system) = attrs.get(semcov::trace::RPC_SYSTEM) {
             data.type_ = Some(rpc_system.into());
@@ -440,8 +449,10 @@ impl<'a> From<SpanAndResource<'a>> for RemoteDependencyData {
     }
 }
 
-impl From<&Event> for ExceptionData {
-    fn from(event: &Event) -> ExceptionData {
+struct EventAndResource<'a>(&'a Event, Option<&'a Resource>);
+
+impl From<EventAndResource<'_>> for ExceptionData {
+    fn from(EventAndResource(event, resource): EventAndResource<'_>) -> Self {
         let mut attrs = attrs_to_map(event.attributes.iter());
         let exception = ExceptionDetails {
             type_name: attrs
@@ -460,13 +471,13 @@ impl From<&Event> for ExceptionData {
             ver: 2,
             exceptions: vec![exception],
             severity_level: None,
-            properties: attrs_map_to_properties(attrs),
+            properties: attrs_map_to_properties(attrs, resource),
         }
     }
 }
 
-impl From<&Event> for EventData {
-    fn from(event: &Event) -> EventData {
+impl From<EventAndResource<'_>> for EventData {
+    fn from(EventAndResource(event, resource): EventAndResource<'_>) -> Self {
         let mut attrs = attrs_to_map(event.attributes.iter());
         EventData {
             ver: 2,
@@ -474,7 +485,7 @@ impl From<&Event> for EventData {
                 .remove(CUSTOM_EVENT_NAME)
                 .map(Into::into)
                 .unwrap_or_else(|| "<no name>".into()),
-            properties: attrs_map_to_properties(attrs),
+            properties: attrs_map_to_properties(attrs, resource),
         }
     }
 }
@@ -484,8 +495,8 @@ impl From<&Event> for EventData {
 /// https://github.com/tokio-rs/tracing/blob/a0126b2e2d465e8e6d514acdf128fcef5b863d27/tracing-opentelemetry/src/subscriber.rs#L839
 const LEVEL: &str = "level";
 
-impl From<&Event> for MessageData {
-    fn from(event: &Event) -> MessageData {
+impl From<EventAndResource<'_>> for MessageData {
+    fn from(EventAndResource(event, resource): EventAndResource<'_>) -> Self {
         let mut attrs = attrs_to_map(event.attributes.iter());
         let severity_level = attrs.get(LEVEL).and_then(|&x| value_to_severity_level(x));
         if severity_level.is_some() {
@@ -499,7 +510,7 @@ impl From<&Event> for MessageData {
             } else {
                 event.name.clone().into_owned().into()
             },
-            properties: attrs_map_to_properties(attrs),
+            properties: attrs_map_to_properties(attrs, resource),
         }
     }
 }